# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trashure API Documentation",
        "description": "Dokumentasi RESTful API untuk Sistem Pengelolaan Sampah dan Penjemputan Trashure.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Trashure API Base URL"
        }
    ],
    "paths": {
...
        "/warga/jenis-sampah": {
            "get": {
                "tags": [
                    "Warga - Jenis Sampah"
                ],
                "summary": "Daftar jenis sampah (Warga)",
                "description": "Mengambil semua daftar jenis sampah yang berstatus aktif, diurutkan berdasarkan nama.",
                "operationId": "18cb9d2c83787243209f5229051e25ca",
                "responses": {
                    "200": {
                        "description": "Data jenis sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data jenis sampah berhasil diambil."
                                        },
                                        "data": {
                                            "description": "Daftar jenis sampah aktif.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/JenisSampah"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/warga/jenis-sampah/{id}": {
            "get": {
                "tags": [
                    "Warga - Jenis Sampah"
                ],
                "summary": "Detail jenis sampah (Warga)",
                "description": "Mengambil detail informasi jenis sampah tertentu berdasarkan ID.",
                "operationId": "044ad38510b76536fd24bef986c4be3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Jenis Sampah.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail jenis sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detail jenis sampah berhasil diambil."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/JenisSampah"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jenis sampah tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Jenis sampah tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JenisSampah": {
                "title": "Jenis Sampah",
                "description": "Entitas katalog jenis sampah.",
                "required": [
                    "jenis_sampah_id",
                    "nama_jenis_sampah",
                    "kategori",
                    "satuan",
                    "status"
                ],
                "properties": {
                    "jenis_sampah_id": {
                        "description": "ID unik jenis sampah.",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nama_jenis_sampah": {
                        "description": "Nama jenis sampah.",
                        "type": "string",
                        "example": "Botol Plastik PET"
                    },
                    "kategori": {
                        "description": "Kategori sampah.",
                        "type": "string",
                        "example": "Plastik"
                    },
                    "satuan": {
                        "description": "Satuan penimbangan.",
                        "type": "string",
                        "example": "kg"
                    },
                    "keterangan": {
                        "description": "Deskripsi tambahan jenis sampah.",
                        "type": "string",
                        "example": "Botol plastik bekas minuman kemasan.",
                        "nullable": true
                    },
                    "icon": {
                        "description": "Nama ikon yang ditampilkan di aplikasi.",
                        "type": "string",
                        "example": "recycle",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status ketersediaan jenis sampah.",
                        "type": "string",
                        "example": "aktif",
                        "enum": [
                            "aktif",
                            "nonaktif"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan token Bearer dari endpoint /api/login untuk mengakses endpoint yang dilindungi.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}